+++ movie-service/src/main/java/com/garret/movies/service/config/mappings/ModelMapperConfigurer.java
@@ -1,4 +1,4 @@
-package com.garret.movies.config.mapper;
+package com.garret.movies.service.config.mappings;
 
 import org.modelmapper.ModelMapper;
