+++ movie-service/src/main/java/com/garret/movies/service/config/mappings/impl/MovieDtoToMovieConfigurer.java
@@ -1,6 +1,6 @@
-package com.garret.movies.config.mapper.impl;
+package com.garret.movies.service.config.mappings.impl;
 
-import com.garret.movies.config.mapper.ModelMapperConfigurer;
+import com.garret.movies.service.config.mappings.ModelMapperConfigurer;
 import com.garret.movies.dao.entity.Movie;
 import com.garret.movies.service.dto.response.MovieDto;
 import org.modelmapper.ModelMapper;
