+++ movie-service/src/main/java/com/garret/movies/service/exception/IncorrectDateException.java
@@ -1,4 +1,4 @@
-package com.garret.movies.exception;
+package com.garret.movies.service.exception;
 
 public class IncorrectDateException extends RuntimeException {
     public IncorrectDateException(String message) {
