+++ movie-service/src/test/java/com/garret/movies/service/util/MovieUtilTest.java
@@ -1,6 +1,6 @@
-package com.garret.movies.util;
+package com.garret.movies.service.util;
 
-import com.garret.movies.exception.IncorrectDateException;
+import com.garret.movies.service.exception.IncorrectDateException;
 import com.garret.movies.service.dto.ActorDto;
 import com.garret.movies.service.dto.CountryDto;
 import com.garret.movies.service.dto.GenreDto;
